refactor(core): split command utilities into focused modules
